// frontend/components/layout/tw-types.ts
export type Space =
  | "space-y-0"
  | "space-y-1"
  | "space-y-2"
  | "space-y-3"
  | "space-y-4"
  | "space-y-5"
  | "space-y-6"
  | "space-y-7"
  | "space-y-8"
  | "space-y-9"
  | "space-y-10"
  | "space-y-11"
  | "space-y-12"
  | "space-y-14"
  | "space-y-16"
  | "space-y-20"
  | "space-y-24"
  | "space-y-28"
  | "space-y-32"
  | "space-y-36"
  | "space-y-40"
  | "space-y-44"
  | "space-y-48"
  | "space-y-52"
  | "space-y-56"
  | "space-y-60"
  | "space-y-64"
  | "space-y-72"
  | "space-y-80"
  | "space-y-96"
  | "space-x-0"
  | "space-x-1"
  | "space-x-2"
  | "space-x-3"
  | "space-x-4"
  | "space-x-5"
  | "space-x-6"
  | "space-x-7"
  | "space-x-8"
  | "space-x-9"
  | "space-x-10"
  | "space-x-11"
  | "space-x-12"
  | "space-x-14"
  | "space-x-16"
  | "space-x-20"
  | "space-x-24"
  | "space-x-28"
  | "space-x-32"
  | "space-x-36"
  | "space-x-40"
  | "space-x-44"
  | "space-x-48"
  | "space-x-52"
  | "space-x-56"
  | "space-x-60"
  | "space-x-64"
  | "space-x-72"
  | "space-x-80"
  | "space-x-96";
export type Gap =
  | "gap-0"
  | "gap-1"
  | "gap-2"
  | "gap-3"
  | "gap-4"
  | "gap-5"
  | "gap-6"
  | "gap-7"
  | "gap-8"
  | "gap-9"
  | "gap-10"
  | "gap-11"
  | "gap-12"
  | "gap-14"
  | "gap-16"
  | "gap-20"
  | "gap-24"
  | "gap-28"
  | "gap-32"
  | "gap-36"
  | "gap-40"
  | "gap-44"
  | "gap-48"
  | "gap-52"
  | "gap-56"
  | "gap-60"
  | "gap-64"
  | "gap-72"
  | "gap-80"
  | "gap-96";
export type JustifyContent =
  | "justify-start"
  | "justify-end"
  | "justify-center"
  | "justify-between"
  | "justify-around"
  | "justify-evenly";
export type AlignItems =
  | "items-start"
  | "items-end"
  | "items-center"
  | "items-baseline"
  | "items-stretch";
export type FlexBasis =
  | "basis-0"
  | "basis-1"
  | "basis-2"
  | "basis-3"
  | "basis-4"
  | "basis-5"
  | "basis-6"
  | "basis-7"
  | "basis-8"
  | "basis-9"
  | "basis-10"
  | "basis-11"
  | "basis-12"
  | "basis-14"
  | "basis-16"
  | "basis-20"
  | "basis-24"
  | "basis-28"
  | "basis-32"
  | "basis-36"
  | "basis-40"
  | "basis-44"
  | "basis-48"
  | "basis-52"
  | "basis-56"
  | "basis-60"
  | "basis-64"
  | "basis-72"
  | "basis-80"
  | "basis-96"
  | "basis-auto";